comicgenerator/requirements.txt
langchain
openai
stability-sdk
pillow
langchain_community